requirements.txt
certifi==2025.4.26
dnspython==2.7.0
pymongo==4.12.1
python-dotenv==1.1.0
